requirements.txt
torch>=2.2
torchvision>=0.17
numpy
pandas
scikit-learn
tqdm
pyyaml
opencv-python
matplotlib
seaborn
grad-cam
